refactored; improve docs
