+++ src/example/di/index.ts
@@ -1,4 +1,4 @@
-import { publisher } from '../../publisher.js'
+import { publish } from '../../publish.js'
 export { CounterView } from './CounterView.js'
 
 export interface SequenceService {
@@ -33,5 +33,5 @@ function* fib() {
   }
 }
 
-const pub = publisher(document.body).create<SequenceService>('service.counter')
+const pub = publish(document.body).create<SequenceService>('service.counter')
 pub.update(new CounterService())
