+++ src/example/progress/ProgressDialog.ts
@@ -1,4 +1,5 @@
-import { Publication, publisher } from '../../publisher.js'
+import { Publication } from '../../publication'
+import { publish } from '../../publish.js'
 import { subscribe } from '../../subscribe.js'
 import { FileService } from './FileService.js'
 
@@ -10,7 +11,7 @@ export class ProgressDialog extends HTMLElement {
   constructor() {
     super()
     this.root = this.attachShadow({mode: 'open'})
-    this.pub = publisher(this.root).create('percent', 0)
+    this.pub = publish(this.root).create('percent', 0)
 
     this.root.innerHTML = `
        <h1>CHUNK DOWNLOAD</h1>
