+++ src/index.test.ts
@@ -1,3 +1,3 @@
 import './example/index.test.js'
-import './publisher.test.js'
+import './publication.test.js'
 import './subscriber.test.js'
